Remove forced EAE local transcoding, change logdir
Discovered that plexmediaserver will fall back to using the transcode
directory (instead of /tmp/pms-*/EasyAudioEncoder) as:
    <TRANSCODE_DIR>/Transcode/Sessions/EasyAudioEncoder

To take advantage of this, the 'plex' user needs to have write access
to /tmp removed. This can be accomplished with:
    setfacl -m u:plex:r /tmp/

With /tmp unable to be written to by the 'plex' user, the default
config for logging filename needs to be somewhere else (this uses
/opt/plex/tmp/prt.log, matching the shared transcode directory in
the wiki).